refactoring

+++ src/modules/registration/registration.service.ts
@@ -1,33 +1,38 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { User } from '../user/entity/user.entity';
 import { TokenService } from '../token/token.service';
 import { ConfigService } from '@nestjs/config';
-import * as bcrypt from 'bcrypt';
 import { TOKEN_TYPE } from '../../shared/types';
 import { UserService } from '../user/user.service';
 import { MailService } from '../mail/mail.service';
+import { CryptService } from '../crypt/crypt.service';
 import {
+  USER_NEED_CONFIRM_EMAIL,
   USER_ALREADY_REGISTER,
   EMAIL_INVALID,
   TOKEN_INVALID,
-  USER_NEED_CONFIRM_EMAIL,
-  BCRYPT_SALT_ROUNDS,
-  SUCCESS,
-} from './const';
+} from '../../shared/const';
 
 @Injectable()
 export class RegistrationService {
   constructor(
-    @InjectRepository(User)
     private readonly configService: ConfigService,
     private readonly tokenService: TokenService,
     private readonly mailService: MailService,
     private readonly userService: UserService,
+    private readonly cryptService: CryptService,
   ) {}
 
-  async register(email: string, password: string, name: string): Promise<any> {
-    const hashedPassword = await this.hashPassword(password);
+  /**
+   * Registers a new user with the provided email, password, and name.
+   * If the user already exists, it throws an error based on the email confirmation status.
+   *
+   * @param {string} email - The email of the new user.
+   * @param {string} password - The password of the new user.
+   * @param {string} name - The name of the new user.
+   * @throws {BadRequestException} - If the user already exists and email is confirmed or needs confirmation.
+   * @returns {Promise<void>}
+   */
+  async register(email: string, password: string, name: string): Promise<void> {
     const user = await this.userService.findUserByEmail(email);
 
     if (user) {
@@ -38,18 +43,24 @@ export class RegistrationService {
       }
     }
 
+    const hashedPassword = await this.cryptService.hashPassword(password);
     const newUser = await this.userService.create(email, hashedPassword, name);
-
-    const token = await this.tokenService.saveUserToken(
+    const token = await this.tokenService.generateUserToken(
       email,
       newUser,
       TOKEN_TYPE.EMAIL_CONFIRMATION,
     );
-
-    return await this.mailService.sendRegisterTokenEmail(user, token);
+    await this.mailService.sendRegisterTokenEmail(newUser, token);
   }
 
-  async confirmRegister(token: string): Promise<{ code: string }> {
+  /**
+   * Confirms the user's registration by validating the provided token.
+   *
+   * @param {string} token - The token for email confirmation.
+   * @throws {BadRequestException} - If the token is invalid or not found.
+   * @returns {Promise<void>}
+   */
+  async confirmRegister(token: string): Promise<void> {
     const user = await this.tokenService.findUserByToken(
       token,
       TOKEN_TYPE.EMAIL_CONFIRMATION,
@@ -57,45 +68,58 @@ export class RegistrationService {
     if (!user) {
       throw new BadRequestException(TOKEN_INVALID);
     }
+
     user.isEmailConfirmed = true;
+    await this.userService.save(user);
     await this.tokenService.removeTokenByUser(
       user,
       TOKEN_TYPE.EMAIL_CONFIRMATION,
     );
-    await this.userService.save(user);
-    return {
-      code: SUCCESS,
-    };
   }
 
+  /**
+   * Initiates the password reset process by generating a reset token for the user with the given email.
+   *
+   * @param {string} email - The email of the user who wants to reset their password.
+   * @throws {BadRequestException} - If the email is not found in the system.
+   * @returns {Promise<void>}
+   */
   async resetPassword(email: string): Promise<void> {
     const user = await this.userService.findUserByEmail(email);
-
     if (!user) {
       throw new BadRequestException(EMAIL_INVALID);
     }
 
-    const token = await this.tokenService.saveUserToken(
+    const token = await this.tokenService.generateUserToken(
       email,
       user,
       TOKEN_TYPE.PASSWORD_RESET,
     );
-
-    return await this.mailService.sendResetPasswordToken(user, token);
+    await this.mailService.sendResetPasswordToken(user, token);
   }
 
-  async confirmResetPassword(value: string, newPassword: string): Promise<any> {
+  /**
+   * Confirms the user's password reset by validating the provided token and setting the new password.
+   *
+   * @param {string} token - The password reset token.
+   * @param {string} newPassword - The new password to be set for the user.
+   * @throws {BadRequestException} - If the token is invalid or not found.
+   * @returns {Promise<void>}
+   */
+  async confirmResetPassword(
+    token: string,
+    newPassword: string,
+  ): Promise<void> {
     const user = await this.tokenService.findUserByToken(
-      value,
+      token,
       TOKEN_TYPE.PASSWORD_RESET,
     );
-    user.password = await this.hashPassword(newPassword);
-    await this.tokenService.removeTokenByUser(user, TOKEN_TYPE.PASSWORD_RESET);
-    await this.userService.save(user);
-    return true;
-  }
+    if (!user) {
+      throw new BadRequestException(TOKEN_INVALID);
+    }
 
-  private async hashPassword(password: string): Promise<string> {
-    return await bcrypt.hash(password, BCRYPT_SALT_ROUNDS);
+    user.password = await this.cryptService.hashPassword(newPassword);
+    await this.userService.save(user);
+    await this.tokenService.removeTokenByUser(user, TOKEN_TYPE.PASSWORD_RESET);
   }
 }
